Add files via upload
Initial check in of log_scan_data and send_transform classes. The log_scan_data class will create a logfile of z elevations broadcast the laser scanner in an ariac simulator. This class also transforms the laserscan type to a pointcloud data type for easier data post processing. 

The send_transform class must be running prior to running log_scan_data. This is a workaround for a bug in the ariac simulation--there is no pre-existing transform between the world and the laser scanner.